requirements.txt
anyio==4.12.1
APScheduler==3.10.4
certifi==2026.2.25
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
psycopg[binary]==3.2.6
python-dotenv==1.2.2
python-telegram-bot==21.6
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
